feat(os-job-runtime): rename job crates and enforce portability spec
