+++ code/packages/rust/Cargo.toml
@@ -171,11 +171,11 @@ members = [
     "sql-csv-source",
     "ean-13",
     "epoll",
-    "job-backend-launchd-files",
-    "job-backend-systemd-files",
-    "job-backend-windows-xml",
-    "job-core",
-    "job-runtime",
+    "macos-job-backend-launchd-files",
+    "linux-job-backend-systemd-files",
+    "windows-job-backend-task-xml",
+    "os-job-core",
+    "os-job-runtime",
     "event-loop",
     "hyperloglog",
     "hash-functions",
