--- code/packages/rust/job-backend-launchd-files/BUILD
@@ -1,1 +1,0 @@
-cargo test -p job-backend-launchd-files -- --nocapture
