--- code/packages/rust/job-backend-launchd-files/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "job-backend-launchd-files"
-version = "0.1.0"
-edition = "2021"
-description = "launchd backend file generation for native job installation"
-
-[dependencies]
-job-core = { path = "../job-core" }
