--- code/packages/rust/job-backend-systemd-files/BUILD
@@ -1,1 +1,0 @@
-cargo test -p job-backend-systemd-files -- --nocapture
