--- code/packages/rust/job-backend-systemd-files/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "job-backend-systemd-files"
-version = "0.1.0"
-edition = "2021"
-description = "systemd user unit backend file generation for native job installation"
-
-[dependencies]
-job-core = { path = "../job-core" }
