--- code/packages/rust/job-backend-windows-xml/BUILD
@@ -1,1 +1,0 @@
-cargo test -p job-backend-windows-xml -- --nocapture
