--- code/packages/rust/job-backend-windows-xml/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "job-backend-windows-xml"
-version = "0.1.0"
-edition = "2021"
-description = "Windows Task Scheduler XML backend for native job installation"
-
-[dependencies]
-job-core = { path = "../job-core" }
