--- code/packages/rust/job-core/BUILD
@@ -1,1 +1,0 @@
-cargo test -p job-core -- --nocapture
