--- code/packages/rust/job-core/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "job-core"
-version = "0.1.0"
-edition = "2021"
-description = "Portable job specification types for native OS schedulers"
-
-[dependencies]
