--- code/packages/rust/job-runtime/BUILD
@@ -1,1 +1,0 @@
-cargo test -p job-runtime -- --nocapture
