+++ code/packages/rust/linux-job-backend-systemd-files/BUILD
@@ -1,0 +1,1 @@
+cargo test -p linux-job-backend-systemd-files -- --nocapture
