+++ code/packages/rust/macos-job-backend-launchd-files/BUILD
@@ -1,0 +1,1 @@
+cargo test -p macos-job-backend-launchd-files -- --nocapture
