+++ code/packages/rust/os-job-core/BUILD
@@ -1,0 +1,1 @@
+cargo test -p os-job-core -- --nocapture
