+++ code/packages/rust/os-job-runtime/BUILD
@@ -1,0 +1,1 @@
+cargo test -p os-job-runtime -- --nocapture
