+++ code/packages/rust/windows-job-backend-task-xml/BUILD
@@ -1,0 +1,1 @@
+cargo test -p windows-job-backend-task-xml -- --nocapture
